Feature/js channel plugin

+++ pymaid/channel.py
@@ -54,7 +54,7 @@ class Channel(object):
                     return
                 self.logger.exception(ex)
                 raise
-            conn = ConnectionClass(sock=peer_socket, server_side=True)
+            conn = ConnectionClass(self, sock=peer_socket, server_side=True)
             bind(conn)
             attach(conn)
 
@@ -85,16 +85,16 @@ class Channel(object):
     def listen(self, address, type_=SOCK_STREAM, backlog=MAX_BACKLOG):
         if self.is_bound:
             self.logger.warn(
-                '%s wants to listen on %s while already bound %s' % (
-                    self.__class__.__name__, address, self.listener.getsockname()
-                )
+                '%s wants to listen on %s while already bound %s',
+                self.__class__.__name__, address, self.listener.getsockname()
             )
             return
         self.is_bound = True
         # not support ipv6 yet
         if isinstance(address, string_types):
             family = AF_UNIX
-            os.unlink(address)
+            if os.path.exists(address):
+                os.unlink(address)
         else:
             family = AF_INET
         sock = realsocket(family, type_)
@@ -108,7 +108,9 @@ class Channel(object):
     def connect(self, address, family=AF_INET, type_=SOCK_STREAM, timeout=None):
         if isinstance(address, string_types):
             family = AF_UNIX
-        conn = self.CONNECTION_CLASS(family=family, type_=type_, server_side=False)
+        conn = self.CONNECTION_CLASS(
+            self, family=family, type_=type_, server_side=False
+        )
         conn.connect(address, timeout)
         self._bind_connection_handler(conn)
         return conn
